requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.2.3
